refactor: services and thunks
