+++ src/pages/profile/ProfileHeader.tsx
@@ -1,0 +1,9 @@
+import { useAppSelector } from "../../hooks";
+
+export default function ProfileHeader(): JSX.Element {
+  return (
+    <section>
+      <img src="https://picsum.photos/800/300" alt="profile header" />
+    </section>
+  );
+}
