+++ src/redux/store.tsx
@@ -1,5 +1,5 @@
 import { configureStore } from "@reduxjs/toolkit";
-import authReducer from "./features/authSlice";
+import authReducer from "./features/Auth/authSlice";
 
 export const store = configureStore({
   reducer: {
